added STCurrency.ts
