+++ packages/swtc-serializer/src/TypesUtils.js
@@ -185,97 +185,6 @@ function Factory(Wallet = WalletFactory()) {
 
   STHash160.id = 17
 
-  // Internal
-  /*
-   * Should handle
-   */
-  var STCurrency = new SerializedType({
-    // Convert the input JSON format data INTO a BYTE array
-    from_json_to_bytes: function(j) {
-      // return (new Currency()).parse_json(j, shouldInterpretSWT);
-
-      var val = new Array(20) // return byte array representing currency code
-      for (var i = 0; i < 20; i++) {
-        val[i] = 0
-      }
-      switch (typeof j) {
-        case "string":
-          // For Tum code with 40 chars, such as
-          // 800000000000000000000000A95EFD7EC3101635
-          // treat as HEX string, convert to the 20 bytes array
-          if (DataCheck.isCustomTum(j)) {
-            val = convertHexToByteArray(j)
-          } else if (DataCheck.isCurrency(j)) {
-            // For Tum code with 3 letters/digits, such as
-            // CNY, USD,
-            // treat
-            //   var currencyCode = j.toUpperCase();
-            var currencyCode = j
-
-            var end = 14
-            var len = currencyCode.length - 1
-            for (var x = len; x >= 0; x--) {
-              val[end - x] = currencyCode.charCodeAt(len - x) & 0xff
-            }
-          } else {
-            // Input not match the naming format
-            // Throw error
-            throw new Error("Input tum code not valid!")
-          }
-
-          break
-        case "number":
-          // TODO, follow the Tum code rules
-          throw new Error("Input tum code not valid!")
-          // if (!isNaN(j)) {
-          //     this.parse_number(j);
-          // }
-          // break;
-        case "object":
-          throw new Error("Input tum code not valid!")
-          // break;
-      }
-
-      return val
-    },
-
-    serialize: function(so, val) {
-      var currencyData = val.to_bytes()
-      if (!currencyData) {
-        throw new Error(
-          "Tried to serialize invalid/unimplemented currency type."
-        )
-      }
-      so.append(currencyData)
-    },
-
-    // Convert the Tum/Currency code from a 20 bytes array
-    // TODO, check the parse value
-    from_bytes: function(j) {
-      if (!Array.isArray(j) || j.length !== 20) {
-        return NaN
-      } else {
-        return new BigInteger([0].concat(j), 256)
-      }
-    },
-
-    parse: function(so) {
-      var bytes = so.read(20)
-      var currency = this.from_bytes(bytes)
-      // XXX Disabled check. Theoretically, the Currency class should support any
-      //     UInt160 value and consider it valid. But it doesn't, so for the
-      //     deserialization to be usable, we need to allow invalid results for now.
-      // if (!currency.is_valid()) {
-      //  throw new Error('Invalid currency: '+convertByteArrayToHex(bytes));
-      // }
-      let s = currency.toString(16).toUpperCase()
-      if (!DataCheck.isCustomTum(s)) {
-        s = convertHexToString(s).replace(/\u0000/g, "")
-      }
-      return s
-    }
-  })
-
   var STVL = (EXPORTS.VariableLength = EXPORTS.VL = new SerializedType({
     serialize: function(so, val) {
       if (typeof val === "string") {
